refactor: 정상 입력을 받는 함수 InputValidator를 다른 클래스로 분리, 파일도 분리
